CameraService: Disable capture message types before stopping preview
If we try to stopPreview in the middle of capture, these
callbacks might occur. In this case, stopPreview will already be
holding mLock while the message handler will try to grab mLock.
Deadlock occurs if the stopPreview and message callback dependencies
are in the same thread.

Issue reproduced with Image Capture/Home Key scenario.

Port from GB: Issue is still reproducible (although extremely
              rarely) in ICS.

Change-Id: I222d00f8bd69c11302a35610095efe0ddc49f876
Signed-off-by: Daniel Levin <dendy@ti.com>