StageFright: Add a 'TimeInterpolator' class
The TimeInterpolator is a mechanism to transform the discreet flow of
media via buffer transfers into a monotonic clock that can be used for
precise scheduling and latency compensation.  While intended to be
used as an "Audio Clock," it can be used in any situation where
discreet data flow can be quantified in terms of time.

Suppose that a thread is reading audio data from a buffer in chunks of
N bytes (which corresponds to T seconds of music).  It is known that
the thread is directly feeding the hardware and that (on average) the
data flow rate is a constant (e.g. because there is a known sample
rate).  However, the time span between reads is random.  If the
time-stamp of the most recent read is used as a "clock," then the
clock will be jumpy and irregular.

The TimeInterpolator can take this jumpy data and convert it to a
smooth and nearly straight line.

Change-Id: Ic86d3534ab7202ae4a5a2fb98607942c550f4ffa
Signed-off-by: Daniel Levin <dendy@ti.com>

[Ported to JB MR2 and squashed with the following changes]

[Daniel Levin <dendy@ti.com>]
    StageFright: Use TimeInterpolator for audio clock.
[Sergii Iegorov <x0155539@ti.com>]
    Prevent negative timestamps propagating to user.
[Gabriel M. Beddingfield <gabrbedd@ti.com>,
 Barry Woodward <a0270655@ti.com>,
 Dmytro Dubovyk <dmytro.dubovyk@ti.com>]
    StageFright: Detect read-pointer mismatch between media TimeInterpolator