Fix Race condition in Stagefright in thumbnail generation
Issue:
Video editor is not able to add CABAC encoded H264 files to project first time.

Observations:
When we add the cabac file to Video editor media server crashes. The crash is
seen in Stagefright library since software avc decoder is not able to decode
the clip.
2 signature of crashes are seen
   1. Buffer book-keeping is crashing in Idle to Loaded state.
   2. Random segmentation fault is happening while freeing the node.

Fix Description:
We have OMXCodec.cpp file containing on_message and stop, two functions which are involved in this race condition.
The variable which is having the race condition in mState.

Steps:
1. We receive error event since decoder is not able to decode the frames.
2. The mState is no in ERROR state
3. Stop function is invoked since the AsyncCompletion semaphore is triggered
4. FTBD and ETB calls are received in on_message while we have mState in ERROR
So we are ignoring the callback and doing nothing due to a condition check at
 the start of on_message function
5. Also we see every FTBD is followed by a ERROR event from codec
6. Now stop function will go ahead in parallel and change the mState to Executing_To_Idle
7. But as explained in step 3 & 4 this state is made ERROR again before the stop function reaches
a critical wait loop for checking transition to happen to LOADED conditon.
8. So we never wait for transition to happen to loaded state since we dont wait on AsynCompletion semaphore
So the component is still transitioning to idle state and we try to free the node. This is crash signature 2
as mentioned above.
9. If we have reached by chance the idle transition in onstatechange function then the buffer count will mismatch.
 the reason for that is explained in step 3. We dont change the ownership of the buffers and ignore them.

Observations with Fix:
1. Thumbnail generation is faster when we have lots of file pushed and one of them has CABAC in it.
2. We can add CABAC H264 file to video editor the first time itself.

Change-Id: I8b2e906261c5fa0c0a08622f9e9c038462f9cd0f
Signed-Off-By: Rajesh Vandanapu<rajeshv@ti.com>
Signed-off-by: Dandawate Saket <dsaket@ti.com>
Signed-off-by: Sergii Doroshenko <x0163979@ti.com>