[VidDec] Change the i/p buffer size if video clip contains bigger samples than WxH
Currently, the i/p buffer size ('nBufferSize') is always set to WxH by Ducati,
and cannot changed from Stagefright since it is a read-only paramter.
So, if video clip contains bigger samples than WxH, the video playback is failed.
This patch checks if the maximum sample size of the video clip is bigger than WxH (i.e.'nBufferSize'),
and allocate the i/p buffer with the maximum sample size if necessary.

Ref: OMAPS00259559

Change-Id: I69b4e7b17107a7a681a8083335317a182eb02e65
Signed-off-by: Daniel Levin <dendy@ti.com>