Workaround for Mpeg2TS grey thumbnail issue
Seek is not supported in Mpeg2TSExtractor. Thus during thumbnail
generation if the software decoder is failing the extractor is not
seeking to the I-frame and is providing the next available frame to
hardware decoder. Thus there is issue in thumbnail generation.
To resolve this, we directly use hardware decoder for all the
Mpeg2Ts Clips.
Now for Mpeg2TS, Mpeg2PS clips by default hardware codec is used
for thumbnail generation.

Change-Id: I526bd9079740a8fb7457fc92b6dce9f82206170d
Signed-off-by: Daniel Levin <dendy@ti.com>