Media: Stagefright: Video playback instrumentation
In ICS the log "video decoder shutdown completed" is currently
a LOGV. This patch changes it to LOGD if debugFps is enabled.
Similarly, it also shows frames dropped if FPS measurement is ON.

The patch is enclosed within #ifdef OMAP_ENHANCEMENT

This patch is required for automation of the video playback
regression test suite.

Redoing this patch since the previous time the patch was not
uploaded correctly to Gerrit.

Change-Id: I8a08141cda2594e400f93c9377e0395fbf9ac9b3
Signed-off-by: Daniel Levin <dendy@ti.com>