Media: Stagefright: Interlaced video playback support
OMXCodec has been modified to handle an event from
Codec with new Index to detect the interlaced stream.
StagefrightMetadataRetriever is modified to configure
Video frame params appropriately for thumbnail generation.
A new color conversion routine has been added for handling
interlaced buffers.

Note# The Changes are done as per OMX1.2 spec, except the below:
According to omx, the interlaced buffer layout needs to be
communicated to IL as part of buffer payload (at the end of
video data).
Since this is a too much of overhead, the implementation is
done based on event handling mechanism between codec and IL client.

Change-Id: I455541696be5d488ce28a1ec737b5945e454a6ad
Signed-off-by: Daniel Levin <dendy@ti.com>