MPEG4Extractor: Handle the haveDeltaTable() function via sub-classing
Change I8924293f14eccf28ad399a156f2f819320aa2541 added a virtual
function to MediaSource which breaks binary compatibility with 3rd
party binaries. For example, this broke WMV playback because the
vtables didn't match.

This is corrected creating a subclass of MediaSource called
MediaSourceHaveDeltaTable and deriving MPEG4Source from the new
subclass. In AwesomePlayer, if the container is MPEG4 then we
statically cast to MediaSourceHaveDeltaTable to use the new function.

Change-Id: Ifb6a43200867ccc64b97923c26ba0eb48f49477e
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>
Signed-off-by: Sergii Doroshenko <x0163979@ti.com>