[AwesomePlayer] Fix for S3D video playback.
S3d mode is not turned on when videos are not played
from the start. That is because the s3d frame layout
information is in the beginning of the video stream.
That is easily visible when you start a s3d movie in
the Gallery, pause it, go to home screen and then
reopen the Gallery again. The paused movie is not in
s3d mode any longer.

A workaround for this situation is to do an extra read
when the first read is from offset different from the
start of the stream. This makes the decoder return
the extra data containing s3d layout information.

Fixes bug OMAPS00270286

Change-Id: I4086e95db6cfa2b231f46e98af98e964db13fe86
Signed-off-by: Ivaylo Todorov<x0177393@ti.com>
Signed-off-by: Devaraj Rangasamy <dev@ti.com>