StageFright: Detect read-pointer mismatch between media TimeInterpolator
When the media content and the TimeInterpolator disagree about where
the read pointer of the audio is then it causes A/V to be out of
sync. This is most likely to happen with a rare media file where the
beginning of the clip is not '0'.

AwesomePlayer now checks for this condition, and if it happens it
will forcibly update the TimeInterpolator's read pointer.

Change-Id: I4c6046c6aa66a521ff59066eed2551176dac2d01
Signed-off-by: Barry Woodward <a0270655@ti.com>
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>