Video Editor: Fix for sending garbage content in Fake BUffers
The Video Editor is sending one frame initially to "test" encoder.
This i/p buffer has no data, it is just allocated from the heap and sent to
So, the data this frame contains is garbage which the encoder sometime fails
within the allocated o/p buffer size.
Changes below does a memset on this buffer before sending to Encoder
Ported #20385 from ICS
Fixes DR: OMAPS00279248

Signed-off-by: Spurthi <spurthi@ti.com>
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>

Change-Id: Iea81345eded3aa2b8598241028094e38e5fc1e96