WFD: Enable manual append SPS/PPS to the first IDR frame.
This reverts commit 35d149a46e9ce3942b7b4b98f0f7f08239babcbf
and adds checking for the first packetized frame
to append SPS/PPS units.

For some reason Ducati provides SPS/PPS buffer, then another one with IDR
frame from beginning, futher IDRs are in the same buffer with their SPS/PPS
units.
But WFD source is expecting ANY IDR prepended SPS/PPS.
So if we met CSD buffer in Converter, we tell the PlaybackSession to check
the next buffer, and if the next buffer's first unit is IDR - tell
the packetizer to append SPS/PPS manually.
This happens once on the session start and checking takes only the first
5 bytes check.

This also fixes the issue with 0:00:00 timestamp which was set to CSD packet.

Change-Id: I71c5bd3db9c9995e6a01f74bb0cdfb850d52c514
Signed-off-by: Illia Ragozin <x0160199@ti.com>