WFD: Stopping SMS without waiting for frame.
On stopping we try to stop the RepeaterSource.
If RepeaterSource is locked by reading from SMS
we can not stop the looper because of onMessageReceived
is locked by SMS->read()
So we can not send stop signal to SMS and unlock read method.
That is deadlock until reading are not unlocked by available frame,
in other words until any change on the display.

We can stop SMS before the stopping the looper.
In that case read will return EOS
and onMessageReceived will never entered again.
So stopping and destroying the looper is safe in this scenario as well.

Change-Id: I76f2858f7647c36786f1426b664aca5d0e666da3
Signed-off-by: Illia Ragozin <x0160199@ti.com>