audio: Multizone: Initial support
Multizone requires explicit route requests that have to be honored
by the AudioPolicyManager. The following functions are introduced
to pass multizone information:

Routing:
- getPrimaryDevices();
- getZoneSupportedDevices();
- setZoneDevices(zone, devices);
- getZoneDevices(zone);
- setSessionZones(session, zones);
- getSessionZones(session);

Volume:
- setZoneVolume(zone, volume);
- getZoneVolume(zone);
- setSessionVolume(session, zones, volume);
- getSessionVolume(session, zone);

The audioSessionId (session parameter above) is used as an unique
identified to a given audio stream. It must be known at the time
an output (from APM) is requested.

Change-Id: I55c79e98559259745b2ac8696729b34194f6cc97
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>