audio: Multizone: More granularity of invalidated tracks
setStreamOutput() invalidates tracks of a given stream type on all
active playback threads. If different tracks are running on different
outputs, invalidating one of them causes gaps on the other tracks and
outputs as they are invalidated too.

setStreamOutput() receives an output handle which wasn't used before,
it's now used to identify the output whose tracks will be invalidated.

Change-Id: Ie23e61f6c5dda8798ef2b2b69468e24677b5ac5f
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>